change env & fix some bug

+++ server/src/main.ts
@@ -1,12 +1,15 @@
 import * as http from "http";
 import fs from "fs";
 
-import { WWW, BLOGS_DIR, ABOUT_ME, FILES_INFO, PORT } from "./env";
+import { ROOT, WWW, BLOGS_DIR, ABOUT_ME, BLOGS_INFO, PORT } from "./env";
 
 // 检查所需文件夹状态
 new Promise<void>((resolve, reject) => {
+  if (!fs.existsSync(ROOT)) {
+    throw new Error(`${ROOT} is it exists.`);
+  }
   if (!fs.existsSync(WWW)) {
-    fs.mkdirSync(WWW, { recursive: true });
+    throw new Error(`WWW: ${WWW} is not exsist.`);
   }
   if (!fs.existsSync(BLOGS_DIR)) {
     fs.mkdirSync(BLOGS_DIR, { recursive: true });
@@ -14,11 +17,13 @@ new Promise<void>((resolve, reject) => {
   if (!fs.existsSync(ABOUT_ME)) {
     fs.mkdirSync(ABOUT_ME, { recursive: true });
   }
-  fs.writeFileSync(
-    FILES_INFO,
-    JSON.stringify({ blogs: [], about: {} }, null, 2),
-    "utf-8"
-  );
+  if (!fs.existsSync(BLOGS_INFO)) {
+    fs.writeFileSync(
+      BLOGS_INFO,
+      JSON.stringify({ blogs: [], about: {} }, null, 2),
+      "utf-8"
+    );
+  }
   resolve();
 })
   .then(() => {
